CHANGE: Converter bug fix.

+++ src/main/java/com/engagepoint/university/messaging/dto/base/BaseDTO.java
@@ -1,11 +1,11 @@
 package com.engagepoint.university.messaging.dto.base;
 
 public class BaseDTO {
-    private Integer id;
-    public Integer getId() {
+    private Long id;
+    public Long getId() {
         return id;
     }
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
